+++ sdk/compute/mgmt/src/main/java/com/azure/management/compute/GalleryImageVersionPublishingProfile.java
@@ -4,41 +4,14 @@
 
 package com.azure.management.compute;
 
-import com.azure.core.annotation.Fluent;
+import com.azure.core.annotation.Immutable;
 import com.azure.core.util.logging.ClientLogger;
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
 
 /** The GalleryImageVersionPublishingProfile model. */
-@Fluent
+@Immutable
 public final class GalleryImageVersionPublishingProfile extends GalleryArtifactPublishingProfileBase {
     @JsonIgnore private final ClientLogger logger = new ClientLogger(GalleryImageVersionPublishingProfile.class);
-
-    /*
-     * The source image from which the Image Version is going to be created.
-     */
-    @JsonProperty(value = "source", required = true)
-    private GalleryArtifactSource source;
-
-    /**
-     * Get the source property: The source image from which the Image Version is going to be created.
-     *
-     * @return the source value.
-     */
-    public GalleryArtifactSource source() {
-        return this.source;
-    }
-
-    /**
-     * Set the source property: The source image from which the Image Version is going to be created.
-     *
-     * @param source the source value to set.
-     * @return the GalleryImageVersionPublishingProfile object itself.
-     */
-    public GalleryImageVersionPublishingProfile withSource(GalleryArtifactSource source) {
-        this.source = source;
-        return this;
-    }
 
     /**
      * Validates the instance.
@@ -48,13 +21,5 @@ public final class GalleryImageVersionPublishingProfile extends GalleryArtifactP
     @Override
     public void validate() {
         super.validate();
-        if (source() == null) {
-            throw logger
-                .logExceptionAsError(
-                    new IllegalArgumentException(
-                        "Missing required property source in model GalleryImageVersionPublishingProfile"));
-        } else {
-            source().validate();
-        }
     }
 }