+++ sdk/servicebus/azure-messaging-servicebus/src/samples/java/com/azure/messaging/servicebus/ReceiveMessageSyncSample.java
@@ -29,25 +29,29 @@ public class ReceiveMessageSyncSample {
         // "<<fully-qualified-namespace>>" will look similar to "{your-namespace}.servicebus.windows.net"
         // "<<queue-name>>" will be the name of the Service Bus queue instance you created
         // inside the Service Bus namespace.
-        ServiceBusReceiverClient receiverClient = new ServiceBusClientBuilder()
+        ServiceBusReceiverClient receiver = new ServiceBusClientBuilder()
             .connectionString(connectionString)
             .receiver()
             .queueName("<<queue-name>>")
             .buildClient();
 
-        final IterableStream<ServiceBusReceivedMessageContext> receivedMessages =
-            receiverClient.receive(5);
+        // Try to receive a set of messages from Service Bus 10 times. A batch of messages are returned when 5 messages
+        // are received, or the operation timeout has elapsed, whichever occurs first.
+        for (int i = 0; i < 10; i++) {
+            final IterableStream<ServiceBusReceivedMessageContext> receivedMessages =
+                receiver.receive(5);
 
-        receivedMessages.stream().forEach(context -> {
-            ServiceBusReceivedMessage message = context.getMessage();
+            receivedMessages.stream().forEach(context -> {
+                ServiceBusReceivedMessage message = context.getMessage();
 
-            System.out.println("Received Message Id: " + message.getMessageId());
-            System.out.println("Received Message: " + new String(message.getBody()));
+                System.out.println("Received Message Id: " + message.getMessageId());
+                System.out.println("Received Message: " + new String(message.getBody()));
 
-            receiverClient.complete(message);
-        });
+                receiver.complete(message);
+            });
+        }
 
         // Close the receiver.
-        receiverClient.close();
+        receiver.close();
     }
 }