+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -1,8 +1,7 @@
 # Release History
 
 ## 1.0.0-beta.3 (Unreleased)
-
-FormTraining Client updates:
+- Fix bug in FormRecognizer API's to support multipage document recognition.
 - Adopt the `training` namespace for Form Recognizer Training Clients
 - Rename parameter `fileSourceUrl` to `trainingFilesUrl` on `beginTraining` method in FormTrainingClients
 - Rename parameter `useLabelFile` to `useTrainingLabels` on `beginTraining` method in FormTrainingClients